better handling of repository selection with local repositories
- update consult-gh.el to The local repository, if any, is now added to the end of history
list when reading a repo name so the user can easily select it by
`M-n. If `consult-gh-prioritize-local-directory is set to `suggest,
the local repo is also used as initial-input so the user does not need
to type it out or use keys to select it. If `consult-gh-prioritize-local-directory is set to
`t, then local repo is used without querying users whenever possible
otherwise the user will be queried as normal.