Improve getting templates when creating a workflow
This commit improves the query to find workflow templates in
`consult-gh-workflow-create`. It adds previews as wellas a new
customizable variable `consult-gh-workflow-template-repo-sources`
which allows users to specify a list of repositories to fetch workflow
templates from. Each source in the list is a string with the full name
of a repository. Workflow YAML files under the ".github/workflows"
path in these repositories can be fetched and used when creating a new
workflow.

This feature enhances the workflow creation experience by allowing
users to maintain a collection of template repositories that contain
their preferred or commonly used GitHub Actions workflows, making it
easier to reuse workflows across different projects.