fix: fix file handling
This commit fixes a bug in the file object type handling by adding
a fallback case for unknown object types.It also renames
`consult-gh-commit-file` to `consult-gh-push-file` in the CHANGELOG
to match the actual function name. It also fixes a bug in file
handling by ensuring proper tempdir paths are used when creating
file buffers.